Added description in Readme.md Bug fixes

+++ ESP_SHD_MotionSensor.h
@@ -8,7 +8,7 @@ class ShdMotionSensor : public ESP_SmartHomeDevice {
 public:
   ShdMotionSensor(uint8_t _pin);
 private:
-  bool handleMqttRequest(char* _topic, unsigned char* _payload, unsigned int _length);
+  bool handleMqttRequest(char* _topic, unsigned char* _payload, uint16_t _length);
   uint8_t pin;
   bool motionDeteced;
   void pinChange();
